requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
Werkzeug==2.0.1
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
six==1.16.0
Werkzeug==2.0.1
python-dateutil==2.8.1
pytz==2021.1